requirements.txt
streamlit>=1.37.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
PyWavelets>=1.6.0

google-generativeai>=0.7.0
